Remove anonymous functions from props.
