+++ client/components/navigation/controller.jsx
@@ -7,6 +7,7 @@ import Header from './header';
 import DrawerContent from './drawer-content';
 import TimeTravel from './time-travel';
 
+import history from '../../config/history';
 import { base } from '../../config/composition';
 import { rules, LEFT_HAND_NAV_WIDTH } from './styles';
 
@@ -24,7 +25,22 @@ class NavigationController extends Component {
   }
 
   render() {
-    const { actions, locations, siteNavOpen } = this.props;
+    const { actions,
+            locations,
+            timeTravelControlsOpen,
+            siteNavOpen } = this.props;
+    const fullScreenParams = {
+      pathname: location.pathname,
+      query: {
+        fullscreen: true,
+        ...location.query
+      }
+    };
+    const onViewFutureAvailabilitiesClick = () => {
+      actions.emitTimeTravelControlsToggle(!timeTravelControlsOpen);
+      actions.emitToggleSiteNav(false);
+    };
+    const onOpenFullscreenClick = () => history.push(fullScreenParams);
 
     return !locations ? null : (
       <div>
@@ -35,7 +51,10 @@ class NavigationController extends Component {
           onRequestChange={actions.emitToggleSiteNav.bind(null, !siteNavOpen)}
           docked={false}
           width={LEFT_HAND_NAV_WIDTH}>
-            <DrawerContent {...this.props}/>
+            <DrawerContent
+              onViewFutureAvailabilitiesClick={onViewFutureAvailabilitiesClick}
+              onOpenFullscreenClick={onOpenFullscreenClick}
+              {...this.props}/>
         </Drawer>
         <TimeTravel {...this.props}/>
       </div>
@@ -46,10 +65,12 @@ class NavigationController extends Component {
 NavigationController.propTypes = {
   deviceWidth: PropTypes.number.isRequired,
   siteNavOpen: PropTypes.bool.isRequired,
+  timeTravelControlsOpen: PropTypes.bool.isRequired,
   timeTravelTime: PropTypes.string,
   timeSliderValue: PropTypes.number,
   actions: PropTypes.shape({
-    emitDeviceWidthUpdate: PropTypes.func.isRequired,
+    emitTimeTravelControlsToggle: PropTypes.func.isRequired,
+    emitToggleSiteNav: PropTypes.func.isRequired,
     emitToggleSiteNav: PropTypes.func.isRequired,
     emitLocationUpdate: PropTypes.func.isRequired,
     emitToggleDisplayLegend: PropTypes.func.isRequired,
